Update: fix the hero ui
