+++ .env.sample
@@ -2,6 +2,6 @@
 GROQ_API_KEY=your_api_key_here
 
 # Google Gemini API Key for ingredient image analysis 
-GOOGLE_GENERATIVE_AI_API_KEY=your_google_gemini_api_key_here
+GOOGLE_GENERATIVE_AI_API_KEY=AIzaSyAbXSGAoo3tcIKc7U9Z81awtzafzhLwY5w
