+++ app/layout.jsx
@@ -1,16 +1,11 @@
-import { Inter } from "next/font/google";
 import "./globals.css";
 import GoogleTranslateWrapper from '@/components/GoogleTranslateWrapper';
+import {Open_Sans, Pacifico} from "next/font/google";
+const kalinaFont = Open_Sans({
+  subsets: ["latin"],
+  weight: ["400"],
+});
 
-const inter = Inter({ subsets: ["latin"] });
-
-const patrickHand = {
-  fontFamily: '"Patrick Hand", cursive',
-  fallback: [
-    '-apple-system', 'BlinkMacSystemFont', 'Segoe UI', 'Roboto', 'Oxygen',
-    'Ubuntu', 'Cantarell', 'sans-serif'
-  ]
-};
 
 export const metadata = {
   title: "Flavor AI",
@@ -27,8 +22,10 @@ export default function RootLayout({ children }) {
         <link rel="preconnect" href="https://fonts.googleapis.com" />
         <link rel="preconnect" href="https://fonts.gstatic.com" crossOrigin="anonymous" />
         <link href="https://fonts.googleapis.com/css2?family=Patrick+Hand&display=swap" rel="stylesheet" />
+        <link href="https://fonts.googleapis.com/css2?family=Material+Symbols+Outlined" rel="stylesheet" />
+
       </head>
-      <body className={inter.className} style={patrickHand}>
+      <body className={kalinaFont.className}>
         <GoogleTranslateWrapper />
         {children}
       </body>
